test: DaoUtils 클래스를 통한 중복 로직 제거
