+++ src/test/java/kitchenpos/dao/JdbcTemplateMenuGroupDaoTest.java
@@ -1,7 +1,10 @@
 package kitchenpos.dao;
 
 import kitchenpos.domain.MenuGroup;
-import org.junit.jupiter.api.*;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Nested;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.TestInstance;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,7 +14,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
-import static org.assertj.core.api.Assertions.*;
+import static org.assertj.core.api.Assertions.assertThat;
 
 @DaoTest
 class JdbcTemplateMenuGroupDaoTest {
@@ -45,16 +48,18 @@ class JdbcTemplateMenuGroupDaoTest {
         private final List<String> MENU_NAMES = List.of("한식", "양식", "중식", "일식");
         private final Map<Long, MenuGroup> savedMenuGroups = saveMenuGroups(MENU_NAMES);
 
-        @Test
-        void list() {
-            final List<MenuGroup> actual = menuGroupDao.findAll();
-            assertEachEquals(actual, asList(savedMenuGroups));
+        private Map<Long, MenuGroup> saveMenuGroups(final List<String> names) {
+            return names.stream()
+                    .map(MenuGroup::new)
+                    .map(menuGroupDao::save)
+                    .collect(Collectors.toMap(MenuGroup::getId, menuGroup -> menuGroup));
         }
 
-        private List<MenuGroup> asList(final Map<Long, MenuGroup> menuGroups) {
-            return menuGroups.values()
-                    .stream()
-                    .collect(Collectors.toUnmodifiableList());
+        @Test
+        void list() {
+            final var actual = menuGroupDao.findAll();
+            final var expected = DaoUtils.asList(savedMenuGroups);
+            DaoUtils.assertAllEquals(actual, expected, JdbcTemplateMenuGroupDaoTest.this::assertEquals);
         }
 
         @Nested
@@ -103,30 +108,8 @@ class JdbcTemplateMenuGroupDaoTest {
         }
     }
 
-    private void assertEachEquals(final List<MenuGroup> actualList, final List<MenuGroup> expectedList) {
-        final var expectedSize = actualList.size();
-        assertThat(expectedList).hasSize(expectedSize);
-
-        for (int i = 0; i < expectedSize; i++) {
-            final var actual = actualList.get(i);
-            final var expected = expectedList.get(i);
-
-            assertEquals(actual, expected);
-        }
-    }
-
     private void assertEquals(final MenuGroup actual, final MenuGroup expected) {
         assertThat(actual.getId()).isEqualTo(expected.getId());
         assertThat(actual.getName()).isEqualTo(expected.getName());
-    }
-
-    private Map<Long, MenuGroup> saveMenuGroups(final List<String> names) {
-        return names.stream()
-                .map(this::saveGroup)
-                .collect(Collectors.toMap(MenuGroup::getId, menuGroup -> menuGroup));    }
-
-    private MenuGroup saveGroup(final String name) {
-        final var menuGroup = new MenuGroup(name);
-        return menuGroupDao.save(menuGroup);
     }
 }
